backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.25.0
motor>=3.3.1
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
httpx>=0.26.0
openai>=1.10.0
google-genai>=0.3.0
groq>=0.4.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
email-validator>=2.1.0
dnspython>=2.4.0
pywebpush==2.3.0
apscheduler==3.11.2
authlib>=1.3.0
itsdangerous>=2.1.0
